chore: Introduce ESLint
ESLint is a fantastic tool for preventing common code issues from being
introduced into the product. This change adds a simple ESLint
integration and does a **first pass* at fixing common issues.

In the interest of sanity, `react/prop-types` has been downgraded to
"warning" from its default of "error". This is because none of the
components today are typed, and fixing the entire app in one PR is
unreasonable.