feat: Introduce @tanstack/react-query
Using the app is a little painful today - every time you navigate to a
new page, the data must be fetched anew. This works _okay_ on a device
on Wi-Fi, but this can be really irritating when out at the crags. Out
there, cell service is much less reliable and can leave one stuck
looking at a loading screen.

Instead, introduce `@tanstack/react-query`, which is a modern hook-based
state management solution for data liveliness. This allows us to show
stale data and refresh it in the background. In the future, it would
allow us to plug in proper offline support, too.

This PR is a **first pass** over some of the pages. Specifically, only
the ones which make `GET` requests without parameters. The rest
shouldn't be a problem to follow the same pattern, but I didn't want to
rewrite the entire app in one PR, and this seemed like a good stopping
point.