feat: Move /areas error from API to UI
If the requested area can't be found (invalid ID, insufficient
permissions, etc), a message should be shown to the end user. This
already happens today. However it does so in the API layer.

This change moves the error-detection (& string selection) up to the UI
layer. In addition to being a bit more "normal", it also lets the
data-fetching be simpler.