feat: Add a responsive navigation header
The current navigation header isn't terribly responsive to device sizes.
It would be nice to work well on small devices (especially with
touch-input) as well as on large devices.

This change makes the top navigation bar responsive and attempts to be
more user-friendly. Specifically:

1. On wide screens, all of the items are shown, with icons and text.
2. On less-wide screens, some of the "secondary" items are collapsed to
   just icons.
3. On even smaller screens, all of the items will be reduced to just
   icons.
4. On very small screens, it will take on a multi-line format, with the
   first line being the home button and search bar, and the rest of the
   items being slotted into however many rows are necessary. Devices of
   this size are generally touch-input devices, so the text is included
   alongside the icon for finger-friendliness.

In the first 3 scenarios, the search box is dynamic and will expand to
fill up any "remaining" room that might exist (due to differences in the
users' devices, such as font size).

Fixes #8