fix: Invalidate the cache after edits
When making an edit to a resource (area, sector, problem), we need to
make sure to invalidate the cache. Otherwise, we run the risk of showing
the user stale data when they go back to the resource's page.

This is a temporary stop-gap -- in the future, we should go all-in on
`useMutation(..)`, which will modify the cache and prevent a refetch at
all. However, that'll take too long, so let's land this first.