fix: Refetch sectors after editing an image
After an image is edited, we send the user back to where they came from
(likely the sector that they were viewing). However, the data that is
shown on the screen is the _previous_ data, living in RAM.

In order to refetch this, we do a very heavy-handed refetch of **all**
sectors. This is necessary because when we're editing an image, we don't
know which sector(s) that image belongs to.

This is better than the bug, but it's not ideal.