fix: Invalidate the cache on mutations
The current data model has lots of interconnectedness - a field which
can be mutated in one mutation can be replicated in several other
queries. This creates a problem where a mutation can inadvertently
result in stale data in the UI.

To address this - without throwing away the benefits of the local
datastore - we introduce a utility to guarantee data consistency:
invalidate (or refetch) everything. This can be overridden
query-by-query in the API, which should let us migrate them one at a
time. The goal here, however, is to ensure that everything remains
healthy.