fix: Odd redirect bug when moving items to trash
When an item (area, sector, problem) was moved to the trash, it kicked
off a race condition which would leave the user in an odd state. The
sequence of events would look like this:

1. User navigates to `/area/123`
2. User starts to edit the area (`/area/edit/123`)
3. User marks the area for deletion
4. User clicks save
5. A `POST` request is sent to trash the item
6. The user is navigated - with in-app navigation - to the home page (`/`).
7. Under the hood, the `/area/123` query was refetched to keep the data
   consistent
8. When the response for loading the data comes in, the API layer
   notices that the server the user should be redirected to
   `https://.../area/123` and **moves the browser** to the new location
9. The new location (`https://.../area/123`) attempts to load the
   now-deleted resource, and hangs.

The redirecting now happens at the component level, rather than at the
API level. This means that the user won't be blindly redirected just
because some possibly-unrelated API call had a side effect.