feat: Centralize permissions-checking
Introduce an `AuthContainer` component to centralize the gate-keeping
logic ("is this user logged in? do they have the right permissions?")
which was previously scattered throughout the app and duplicated onto
each page.

As a bonus, this makes it easier to see which routes need which
permissions, by looking at the static declarations in `AppRoutes`.