Don't open problems in a new window/tab
Our telemetry shows that the slowest page (on average, 95th percentile)
to laod is `/problem/*`. Virtually all of the sessions that are slowest
begin with a pageload event -- someone is navigating directly to
`/problem/:problemId`.

One possibility is that they're being sent links or clicking on
bookmarks. This is possible (and, indeed, probable).

Another possibility to enter this flow is from the `/problems` page.
From here, navigating to any problem opens a new tab -- which requires
reloading all of the network-fetched resources. Static assets like CSS,
HTML, and JS are cached. However, the lazily-loaded content is not
fetched.

This change makes the problems default to opening in the same tab. With
the new data layer, this makes navigation forward and backward much less
costly. If someone wishes to open a problem in a new tab, they can
easily do so with whatever mechanism their browser provides.