Do a lazy-load with fallback
When the app is built, it's split into several different chunks, in
order to reduce initial load time (downloading and parsing ≥ 3MB of
JavaScript can be very taxing for some networks and browsers). However,
this creates issues when deploying new versions, as these chunk names
are hashed and change for almost every build.

This means that if a user is using the app, and the CI/CD process
deploys a new version, they might be hit with an error screen when they
browse to a new page. This is very uncool.

What this change does is watch the chunk-loading process and, if that
fails, reloads the page. This should be enough to allow the user to
recover, as they'll get a fresh set of chunk-pointers loaded when the
newly-deployed app is loaded. It's not a _great_ experience, but is
hopefully better than crashes.

 ## Testing

To test that this works, I did the following:

1. `yarn build` to generate a production bundle
2. `npx serve build` to get a local server up and running
3. Navigate to <http://localhost:3000> to load the app
4. In another terminal, `rm build/static/chunk-about-*` to delete the
   chunk for the "About" page
5. Click "About" in the header

At this point, the app will:

1. Attempt to load the `/static/chunk-about-*` resource
2. Fail to load that resource
3. Put a marker into `sessionStorage`
4. Reload the page
   1. (NOTE: you'll be left on an error page because the `serve` package
      doesn't do SPA routing correctly)

After, I:

1. Went to <http://localhost:3000> to get the app loaded again
2. Clicked "About" again

At this point, the app will:

1. Attempt to laod the `/static/chunk-about-*` resource
2. Fail to load that resource (again)
3. See that we've failed to load that previously
4. Show an error message (to prevent a refresh-loop)