Rework the trash page
Update the trash page (`/trash`) to use the mutations & queries pattern.
This is a rather pointless change, but it provided an opportunity to fix
up the typing around `usePostData(..)` and `useData(..)`, which was a
bit hodge-podge.

This also fixes a small error wherein we trigger fetches of
known-deleted resources when a problem is moved to the trash (we should
apply this pattern to other media types, too).