Combine /problems and /filter
This is a substantial change which overhauls the `/problems` and
`/filter` endpoints. The short version is that it removes `/filter` and
replaces it with client-side filtration of the data.

The advantages to this is that it happens client-side, so it's possible
to load the problems list once and then do the filtering offline. A side
advantage is that it removes a somewhat-redundant item in the
already-crowded navigation bar.

One disadvantage is that it happens client-side -- as the data set
grows, peformance might become an issue (on phones in particular). This
change includes performance monitoring of the critical path so we can
keep an eye on this and see if we need to revisit this approach. This
change was tested against the largest dataset (buldreinfo.com) and it
performed adequately, even with a 6x CPU throttle.

Finally, this change introduces some client-side caching of the
`/problems` data so that once the page is laoded, future visits to that
page should show up nearly instantly (with the data refreshed
asynchronously).