Don't crash on failing profile data
It's possible for the profile data to fail to load - I don't know how,
but we have at least one crash in the wild that happened. If this
happens, let's record the failure (so that we can investigate) and show
the user a friendly error message instead of crashing the app.