Fix an issue with the filter not always applying
On initial load, the filter won't filter out based on grade levels. This
is because the grade information is only set when the `<GradeSelect />`
component is rendered .. but it's not rendered by default.

Fix this by listening for the grade information to change and always
keeping the state updated.

This also introduces a `useGrades()` hook to make this information
easier throughout the app (though this cleanup has not yet been
performed).