Rework the /user/<uid>/settings page
The page which lets you control which regions to show under the same
site was a bit annoying to use - it would reload every time something
was changed, making it really flashy and awkward.

This is no longer necessary, as we can reload the data in the
background. However, this change goes one step further and directly
modifies the local datastore to avoid an extra round-trip -- mainly just
for fun and practice :)