Allow filter to scope to specific areas
A common use case for me is to use the filter feature when I'm already
at an area and I want to find something to climb (decision paralysis).
When I'm doing this, it's annoying to have so much stuff on my little
phone screen - and most if it adds zero value because I'm *already at an
area*.

This feature allows the user to scope their filtration to specific
areas.

It also allows sections of the form to be reset individually.