Restore search partial-query persistence
When searching, often you get good enough results without typing the
whole query. In this case, you want to be able to click on a result and
go there. However, sometimes you want to go to another result.

It would be great to have the search box keep the previously-typed query
so that the results are already preloaded for rapid navigation.

This change does exactly that.