Don't use indexes as keys
When generating a list of things, each item needs a key to give React a
handle for internal bookkeeping. This should be based on the item being
rendered so that if the data changes (fetching, reordering, editing,
etc) then React can correctly identify which item(s) need to be
rerendered.

Prior to this change, it was relatively common to use the index of the
item in its respective list as the key. This seems to work, but if a new
item is inserted, React won't know that `0` !== `0` anymore, and things
will get out of sync.

This PR makes it an error to use the index as part of the key
generation, and does a pass through the codebase to attempt to generate
a useful key for each item. This is a best-effort fix and seems to pass
the smoke-test. However, I'm sure that there are cases which were missed
(due to a lack of typing). This will not cause any major problems for
users, and will gradually reveal itself as more and more of the codebase
gets typed.