Encforce typing of useData(..)
This changes the default type of `useData(..)` from `any` to `unknown`.
The implication of this is that typing isn't _required_ to initiate a
request, but a type definition is now required to _use_ the result of
the request.

This small change should introduce a "pit of success" for future queries
by needing their types provided from the get-go.