chore: Introduce Storybook
[Storybook] is a fantastically-useful utility which allows components to
be rendered and tested in isolation. This makes it possible inject
problematic data, force temporary states (eg, loading), and cover
hard-to-remember scenarios.

This change introduces Storybook to the repo, and adds some very basic
stories to demonstrate that it's functional. Stories for the more
complicated components will come in future PRs in order to keep this
practical to review.

[storybook]: https://storybook.js.org/