Collapse the area list with many areas
When filtering for a problem / route, it can be overwhelming (to the UI)
if the user has many (or all) of the areas enabled on their profile. In
cases like these, put a cap on the maximum size of the areas shown -- by
default. If the user wishes to, they're able to expand (and then
collapse) the area list to avoid nested scrolling.