Restore the map on the filter page
Make it easier to get a sense of where the different problems and
sectors are when filtering the routes/problems by showing a
live-updating map on the filter page.

This map will be off by default on mobile devices, as it might be too
cluttered and/or battery-intensive and/or network-intensive. However, it
can be easily enabled with one button press.