chore: Centralize redirect behavior
The current redirect behavior is very endpoint-specific. This is a bit
annoying to work with, and susceptible to errors.

Instead, have `useData(..)` check for the redirect logic and see if we
should be somewhere else. This adds a `redirectUi` element to the
returned object from `useData(..)` which the caller can choose to return
if they'd like to (but the client will be redirected regardless).