Refactor the problem reordering feature
The `SectorEdit` page has a component to allow the problems for a sector
to be presented in a different order. This was implemented in a way that
polluted the state with unexpected props (it just made use of
JavaScript's type flexibility).

This makes it a bit annoying to type the rest of the component, so this
change pulls it out to a separate component and reworks the logic a
little bit to add type safety.