Introduce a responsive filter page
The current filter page has a lot of information, and it's presented in
the same way on all kinds of devices -- no matter how big the screen is.

This change introduces some CSS-powered responsiveness to the page. On
large screens (laptops, desktops, possibly landscape tablets),
better-utilize the screen real estate with a two-column layout. On
phones, however, keep the stacked layout (and hide the filter by
default).