chore: Split up api.ts
`api.ts` contained a whole lot of things -- in addition to being hard to
navigate, it made it unclear what was being used where. This change
splits up `api.ts` into a little package which only exports the
necessary functions. It also makes it clearer which operations haven't
been migrated to the data-fetching model.