Spruce up the elevation display
Do a little revamping of the display of the elevation that's under the
current cursor when editing a sector. This is admittedly a bit of
overkill, but it at least hides the debounce state logic from the
component level and exposes a cleaner API.

By utilizing the `@tanstack/react-query` library, we can expose more
information about the underlying data (ie, is it being fetched or not).

And, on the off chance that the user ever requested the _exact same
latitude & longitude_, it'd be filled instantly from cache!