Extract data-fetching from SvgEdit
`SvgEdit` was the last component that used `fetch`-based data fetching;
this PR updates the process to reuse `useProblem(..)` to utilize the
data-fetching library that is used.

This doesn't do anything to de-complicate the component state (and makes
the process a bit _less_ readable in some ways), but that can come in a
future change.