Remove accessToken from file-download query params
Access tokens [very explicitly] should NOT be passed in query strings.
Per the OAuth bearer token RFC:

> Don't pass bearer tokens in page URLs:  Bearer tokens SHOULD NOT be
> passed in page URLs (for example as query string parameters).
> Instead, bearer tokens SHOULD be passed in HTTP message headers or
> message bodies for which confidentiality measures are taken.
> Browsers, web servers, and other software may not adequately
> secure URLs in the browser history, web server logs, and other
> data structures.  If bearer tokens are passed in page URLs,
> attackers might be able to steal them from the history data, logs,
> or other unsecured locations.

The app _mostly_ follows that rule, except when downloading the PDFs for
the areas/sectors/problems. These resources need the access token
because some resources are restricted. However, they don't need to be
passed with the query string.

Instead of passing them in the query string, perform a GET request and
pass them as headers, like we do with other requests in the app. The
download links are still rendered as normal links, so if the user
middle-clicks on it (for example) then they'll get an _unauthenticated_
version served to them. This is not perfect, but it's the best that we
can do, unless we moved the bearer token to a cookie or something.