feat: Use a RangeSlider for FA year filtering
The first ascent year filter control is just like the grades: a ranged
value. This change follows the convention already introduced by the
grades, and uses a `RangeSlider` for the input, instead of separate
drop-down inputs.

There's _possibly_ an opportunity for code reuse, but it gets a bit
tricky -- while the controls look visually similar, they have different
behaviors internally. Combining them might make sense, but it also might
not. I left a comment in there for a future explorer to consider on a
rainy day :)