chore: Fix some strict-mode TypeScript errors
We (currently) don't use TypeScript in strict mode, which results in a
lot of _probably_-okay errors slipping through. However, it leads to
more of a development struggle as static analysis is harder.

For this change, I uncommented `// strict: true,` in the `tsconfig.json`
file and started fixing errors. For this pass, I focused on the errors
in the general vicinity of the `/problems` route.

More to come in other PRs :)