feat: Introduce state into the filter URL
It's really great that the state of the filter on `/problems` is saved
to the browser-side local storage. However, it can be a bit clunky with
the limitation of having just one configuration -- and also with it
being private.

Two scenarios which this change aims to address:

 1. Planning for an upcoming trip.
    - When discussing with a group ideas for where we could go climbing
      tomorrow or this weekend, it can be useful to ensure that everyone
      is looking at the same data.
 2. Saving multiple filter configurations.
    - It can be nice to have different filters saved for myself. For
      example, one might want to have ready access to:
      - I want to go sport climbing tomorrow, anywhere
      - I want to go to _Location X_ tomorrow, any type of climbing
      - Walls that will be shady in the afternoon

This change addresses both of those use cases by taking the
current filter, encoding it into a string, and then stuffing it into the
hash fragment of the URL.

The user can then take their current URL and send it to their friends:
> What do you think about these routes? <URL>

Similarly, the user can also bookmark the current URL in their browser's
bookmarks feature. Then they can easily hop between different filters.

Future expansions could include moving this to the server-side of things
to allow for _even better_ favoriting / sharing / syncing. However, this
is a relatively cheap way to test the popularity of the feature before
committing to a larger build-out.