fix: Handle edge cases for SunOnWall & Sunrise
The "sun-on-wall" and "sunrise/sunset" widgets would have questionable
(and incorrect) behavior during exceptional times: midnight sun (and
presumably polar night) in the Arctic Circle.

This was previously handled, but a regression was introduced in 76e135f

This change also unbreaks the Storybook setup - when it was bumped to
Storybook v8, the Storybook v7 configuration files weren't completely
migrated. Maybe we can add some safety around this in a future PR to
prevent regressions.