fix: Correct displayed route count
Conditional rendering with `{foo && <Bar />}` works great -- as long as
`foo` is _guaranteed_ to be either `null`, `false`, or `undefined`.
However, once you throw in a number, it gets complicated:

```ts
false && "foo"
```

React will see this as "render `false`" and render nothing. However:

```ts
0 && "foo"
```

This will render a JSX node of `0`, instead of nothing.

This change does a cleanup of the `Area` component, as the bug was
observed there:

screenshot

I'll look to see if there's an ESLint rule we can add to prevent this
pattern in the future, rather than having to know to not use this
shorthand.