chore: Rework SvgEdit & related utils
This change started as an attack on the strict mode violations related
to `SvgEdit` (the highest number in a single file, but also the related
cluster). In the process, much of the inner mechanics of the `SvgEdit`
component were reworked, with a focus on compartmentalization (for
maintainability / extensibility) and performance.

There is no observable feature difference in my (limited?) testing.
However, I am willing to believe that there are some bugs and/or
features inadvertently dropped.

This change also pulls `SvgEdit` into Storybook, so that it can be
tested in isolation - including by devs who might not have permissions
to create all of the necessary data to test in production. This is a
pattern which I would like to repeat in future components.