feat: Improve filter/group/order options on problem list
When looking at a sector (or area), the current experience is fairly
lacking. A common scenario is when I'm already at a particular crag, and
want to move to a different wall. Today, my options are to clumsily use
the list of problems on the `/area/:areaId` page, or use the `/problems`
page with some aggressive filtering.

This change aims to merge the two experiences. A subset of the filtering
options (grades, types) are brought into the problem list, and the
existing functionality is enhanced. The main enhancement is additional
group-by options: area, sector, type, rock, and none. (Note that
superfluous filter types are hidden when they aren't applicable).